Make Proxy work with decode_data=False
In a StackOverflow post [1] a user is trying to use Proxy to create an
open proxy. However, the current implementation of Proxy requires that
`decode_data=True` be set when instantiating `aiosmtpd.smtp.SMTP` which
in turn requires [2] subclassing Controller as done in `UTF8Controller`
in `aiosmtpd/tests/test_handlers.py`.

The Proxy handler should be usable without subclassing Controller, so it
should work with `decode_data=False` (the default). This patch makes it
work with either by checking the type of `envelope.content`.
The module-level constants `EMPTYSTRING` and `CRLF` are only used in
Proxy, so they are removed and replaced with bytes-equivalents
`bEMPTYSTRING` and `bCRLF`, and `bNLCRE` is added as an analogue to
`email.feedparser.NLCRE`.

[1]: https://stackoverflow.com/q/43278998/1570972
[2]: https://stackoverflow.com/a/43639097/1570972