Install as "aiosmtpd" instead of "smtpd"
The $PATH/smtpd command is contested by multiple popular SMTP server
implementations: Postfix installs to /usr/lib/postfix/sbin/smtpd on
Ubuntu, and opensmtpd installs to /usr/sbin/smtpd. If aiosmtpd tries to
install the "smtpd" command upstream, it will probably be changed by
most distributions. If we pick a less contested command name, it is more
likely to be available uniformly across the distributions that choose to
package aiosmtpd.