Handle long lines
When an aiosmtpd server receives a line that is long enough to overflow
the internal buffer of StreamReader (either when reading a command line
or a DATA line), StreamReader.readline() raises a ValueError.

Use StreamReader.readuntil() instead of StreamReader.readline() to give
the proper error responses when long lines are sent by the client.