Prepare for use under socket activation
systemd-style socket-activation is a nice safe way to launch a process
that is already unprivileged.

It works by the parent process opening a socket, and then doing its
own privilege dropping.  This means the child process starts with an
already-listening file descriptor, but no elevated privileges.

To make use of this, the user needs to be able to pass an existing
socket to the aiosmtpd controller.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>