Remove TestLoop test case
Good question in the (recently deceased) `TestLoop` test case:

> is it even worthwhile?

Answer: NO, IT IS NOT.

The principle of unit testing is to observe (and verify) the EXTERNALLY
OBSERVABLE BEHAVIOR of the units under test. It SHOULD NOT concern itself
with how the behavior is achieved.

`TestLoop` totally violates this principle; it concerns itself too much
with how INTERNAL mechanisms should work.

In addition, every single test in `TestLoop` raises the warning:

`RuntimeWarning: coroutine 'AsyncMockMixin._mock_call' was never awaited`

> it does give us coverage.

Taking out the whole `TestLoop` test case resulted in coverage dropping by
a MERE TWO LINES. To make it worse, those two lines are simple lines whose
logic is unambiguous.

Totally not worth the effort.

So I ripped out the whole `TestLoop` class, and stick a pragma:nocover
on the two lines (actually, the first line in the `except` block).