Centralize error handling for HELO/EHLO & AUTH needed
They return (almost) the same responses. It's better to centralize them so if
we need to change them all at the same time, we can do it in one place.

(Also, they are called in the code > 3 times each. So, it fulfills Martin
Fowler's "Rule of Three" guidelines for code duplication.)