Change test_handlers.py to use get_controller
"Naked" Controller should be used only when we're actually testing the
Controller implementation (in test_server.py -- not touched in this
commit)

Because of some requirements to provide non-default port in
test_handlers.py, we need to modify the get_controller fixture as
well.

While we're at it, we also add some docstring for documentation.