Enable dynamic determination of X-Peer address
Previously the X-Peer address was hardcoded as "::1".

This potentially causes test failure on systems with IPv6 disabled,
because `transport.get_extra_info('peername')` will return 127.0.0.1
instead of ::1

So rather than hardcoding the X-Peer address, we fetch the actual
X-Peer address from the Proxy's session instead.

In the process we also change how "source" and "expected" is built, to
ensure that the two are always in-sync with each other.