Move generation of XML into "cov" part
DoubleDurnit, I was so spoiled by pytest's one-shot generation of XML
and HTMLCOV that I forgot this is still the old way of doing things.

It's not that the XML file is "wrongly" named, it's that the XML file
was generated by "diffcov" which we don't run at all!

Travis, be invisible.

[skip travis]