test_handlers.py: Improve FakeParser implementation
Rather than a fixture that merely does instantiation of FakeParser, now
the fixture is more sophisticated and perform not only instantiation of
FakeParser, but also catching exceptions and results.

All leveraging the mighty useful Protol and SimpleNamespace types.

As a side benefit, the test cases look more uniform since there is no
longer any need to wrap statements in "with pytest.raises():" anymore.