More granular process
Split the server Future creation into two steps, and type hint each
step.

Originally went for an `assert isinstance`, but a bug in Python 3.6's
asyncio debug mode made that assertation fail horribly.

That is why we use two-step assignments there: The first assignment is
properly hinted and the second assignment is to record the result.