Cleanup of Comments & Annotations
Also moving out some attributes-initialized-as-None to class level. This
serves two purposes: (1) Make __init__ tidier, and (2) Allow not using
Optional (although it proves to be useful only for one attribute).

Don't worry about different Controller instances sharing same
attributes; once the attribute is assigned to using self.<attr>, it
ceases to be a class attribute and becomes an instance attribute.