Remove "alpha" designation from NEWS.rest
"alpha" should indicate only "stages" of implementation as we get nearer
to release.

NEWS.rst should instead show/summarize _all_ the new (and good) things
in release. Or, in other words, a sum total of all the "alpha"s and
"beta"s and "rc"s prior to release.

"aiosmtpd-next" stays because it's the 'date' of release. Kind of like
"HEAD" in git.