Update comment on the GA yml file
I originally use the term "stable", which was NOT used when I first
created this file. Then Ned Deily put in the definition for "stable" --
which, for the record, I am actually disagreeing with -- so I had to
coin a new, different term.

I chose "frozen" to reflect the fact that the binaries won't ever be
officially updated again.