Don't Edit User-Supplied TLS Context (#196)
* Take out code that edits tls_context

"We are all responsible & consenting adults."

If user passes in a tls_context that's ... 'unusual', we shouldn't modify
it. Oh yes, we don't like it, we'll express our dissatisfaction, but we'll
let the user shoot themselves in the foot.

In summary, the changes are:

* REMOVE: Code that edited tls_context
* ADD: log.warning if tls_context is ... 'unusual'
* ADD: Test Cases for new __init__ behavior
* ADD: log.warning on auth_required without auth_require_tls
* CHANGE: "Rule of Three" D.R.Y. applied to some stuff in the test cases

* Update NEWS with CHANGES

* Grab tox.ini & housekeep.py from newer PRs

* Change from backslash to parens line continuation

* Correct the warning message

* Add test cases

* Add comment

* Remove test cases similar to the ones in test_starttls

Also *greatly* simplify tests still in test_smtp.py by no longer
bothering to start/stop a Controller.

And in the process, remove a no longer used class.

* Simplify some test cases in test_smtp.py

By removing usage of ExitStack and go with @patch + SMTP client context
manager directly.

* Silence some `qa` protests

* Convert some ExitStack to @patch

Mostly to silence PyCharm's code inspection that complains of patch not
fulfilling enter_context's signature.

But the side effect is cleaner, simpler code.

* Make housekeep.py use colorama if available

It's a Sphinx dep so there is a chance that it's pulled in.

* Suppress messy logging output

* Simplify coverage exclusion

* Simplify tox directory structure

Also potentially speed up repeated tests (as long as you don't engage
the "housekeep.py superclean" process...)

* Add envdir for pypy3

* Don't do py[co]/pycache deletion from cleanup step

Because now tox's envs for same Python version share the same venv.

Hopefully with the removal of pycache_clean, the py[co] files and the
__pycache__ dirs  will be kept and reused between testenvs, providing a
slight speedup.

* Improve housekeep.py argparsing

* Adds "--force" option
* Automatic gathering of available actions
* Informative output about actions it will take

Also: rename "cleanup" to "gather"

* Use shared envdir for "py" testenv

GitHub Actions recommend usage of "py" testenv rather than specifying
the exact version of Python used (see unit-testing-and-coverage.yml).
This will result in tox using system's Python version, which in case of
GitHub Actions will be provided as specified in the workflow YML.

Because "qa" and "docs" likewise will use system's Python, we merge all
three into one envdir, hopefully shaving some seconds off GA testing.

* Get improved housekeep.py and tox.ini

from the "command_call_limit" branch.

* Fix "instance attrib outside init" error

* Suppress warnings before controller start

If not then warnings already got out of the door resulting in
particularly noisy and annoying console clutter.