Rewrite assert_auth_* helper funcs
Now it compares the tuple of (code, message) against the expected tuple.

This is done so when test fails, one can quickly see the _whole_ status
sent by the server, because oftentimes information vital to
troubleshooting is contained in the message portion.

Also, reorder the test because apparrently TestCase.assertEqual() treats
the FIRST arg as the EXPECTED value, while the SECOND arg is the ACTUAL
value.