Encode challenge as utf8
For characters <= \x7F, encode() and encode("ascii") will be identical
anyways. But if the challenge str contains ... 'exotic' characters (for
example, emojis), then encode("ascii") will fail while encode() will
always succeeds.

Whether client will be able to understand the challenge or not, that
depends on the client, not our concern.