Add optional and default disabled logging for AUTH
Now that we have sanitized the main logging, I realize that logging the
actual AUTH interaction _might_ be useful, at least during development.

So I added a knob to enable logging of AUTH interaction, defaulting to
False (disabled), with a stern warning in the docstring, and some
warn()s in the code itself.