BUGFIX: RFC4954 behavior for lone "=" (#232)
* Remove special handling for "=" during AUTH

My mistake, I misunderstood the meaning of lone "=" in AUTH/SASL.

It represents an empty bytes b"", which is actually the simple Base64
decoding of "=". Or, in other words:

    b64decode(b"=") == b""

"=" does NOT mean that username = password = b""

* Adapt Test Cases to removal of "=" special handling

* Rewrite assert_auth_* helper funcs

Now it compares the tuple of (code, message) against the expected tuple.

This is done so when test fails, one can quickly see the _whole_ status
sent by the server, because oftentimes information vital to
troubleshooting is contained in the message portion.

Also, reorder the test because apparrently TestCase.assertEqual() treats
the FIRST arg as the EXPECTED value, while the SECOND arg is the ACTUAL
value.

* Remove handling for login is None

login cannot be None at those points; it can be b"" (empty), but can NOT
be None.

Similarly for password

* Update NEWS.rst

* Update handler.rst

_auth_interact is being deprecated and a more powerful challenge_auth
has been implemented.

* Restructure to silence coverage protests

* Exclude from coverage because it's being deprecated

* Remove leftover misunderstanding

* Encode challenge as utf8

For characters <= \x7F, encode() and encode("ascii") will be identical
anyways. But if the challenge str contains ... 'exotic' characters (for
example, emojis), then encode("ascii") will fail while encode() will
always succeeds.

Whether client will be able to understand the challenge or not, that
depends on the client, not our concern.

* Add logging (DEBUG) of challenge sent

Also add another reference to another place in RFC4954 that explicitly
requires trailing space after 334