Implement sanitized_log function
As suggested by @waynew

The issue is to bring forward, emphasize that some log lines need to be
sanitized.

Calling log.func(..., sanitized(...)) works but the sanitize() function
gets shoved to the right, possibly causing future maintainers to miss it
and thus not realizing that a sanitizing is needed on that point.

By replacing the leftmost func call with a kind-of-wrapper function, it
is hoped that future maintainers, when editing these lines, won't miss
the need to sanitize the logging message.