Add testing.rst
testing.rst contains documentation for pytest plugins, fixtures, and
markers used in the test suite.

The documentation is built using sphinx-autofixture, so that dependency
is also added to tox.ini, requirements.txt, and
unit-testing-and-coverage.yml.

This led to some major rewrite of docstrings within conftest.py to
maintain a "Single Source of Truth" principle, as it then gets pulled by
 the "autofixture" directive in testing.rst

In addition, "client_data" marker is also added to the pyproject.toml
file as an option for pytest.