Fix LGTM Alert
Damn you LGTM ... that 'missing' __eq__ was intentional! Why are you not
considering the "suppression" to be "this is intentional. Please do not
bug me about this"?

Anyways, I was forced to write __eq__ which does nothing but pass the
call to dict.__eq__.

And then I have to write a test for that because coverage dropped.
Aaaugh!

Huh.