Add return "::1"
Mainly to silence pytype error.

We never expect the connection to succeed (in fact, it should NEVER
succeed), but pytype doesn't think so, and thus the non-exception branch
will (in the impossible case) return None.

Now it will return "::1" instead.

Also added a test case to simulate this unlikely scenario.