Change equality test to membership
Apparently EAFNOSUPPORT is also a possible socket error if IPv6 is not
available.

Rather than listing _everything_ in a series of equality tests, we grab
the constants and stick them in the attributes of the `IP6_IS` class.

This also enables "on-the-fly patching" by changing the values of the
IP6_IS attributes.

Also update docs.