Change shared ctypes to MP.Queue
PyPy has problems with shared ctypes Value, especially on Windows. Not
sure why. This resulted in flakey pypy3-* tests.

Converting to MP.Queue might be slightly slower, but it's much more
stable/deterministic. And since this change impacts only 3 tests out of
several hundreds, it's worth it for the expansion of Windows test scope.