Allow setting port=0 to use OS-assigned port.
Using port=0 is useful for ephemeral servers (eg for tests) where a
given port can't necessarily be guaranteed to be available.

Setting port=0 was failing because the _trigger_server method then
attempted to connect to the server with port 0, which didn't work.

Fix this by updating the port from the server's socket before attempting
to connect.