Use trustme for test TLS certificate and key
With the current approach TLS certificate and key are hardcoded in
separate files and distributed as part of wheels.
This is unnecessary for end users of the package, also this brings
difficulties while supporting Python 3.13, see:
https://github.com/aio-libs/aiosmtpd/pull/473

Use trustme for generating TLS certificate and key on the fly, as it is
done in other aio-libs packages.

Initially proposed in:
https://github.com/aio-libs/aiosmtpd/pull/473#discussion_r1745661940