smtp.py passing mypy
This brings up a bit of a discussion point: the way I have it now, the
new DATA_CHUNK hook returns Optional[bytes]. The idea is that it will
typically return None if last=False unless there was an early-return
error but the final call with last=True is always expected to return a
value. This can't exactly be expressed with annotations so I needed
some asserts to pass mypy.